actualizacion administracion de usuarios

+++ src/main/java/com/liderbs/dataaccess/dao/IUsersDAO.java
@@ -3,12 +3,11 @@ package com.liderbs.dataaccess.dao;
 import com.liderbs.dataaccess.api.Dao;
 
 import com.liderbs.modelo.Users;
-import com.liderbs.modelo.UsersId;
 
 
 /**
 * Interface for   UsersDAO.
 *
 */
-public interface IUsersDAO extends Dao<Users, Long> {
+public interface IUsersDAO extends Dao<Users, Integer> {
 }
