+++ src/main/java/com/liderbs/dataaccess/dao/UsersDAO.java
@@ -3,7 +3,6 @@ package com.liderbs.dataaccess.dao;
 import com.liderbs.dataaccess.api.HibernateDaoImpl;
 
 import com.liderbs.modelo.Users;
-import com.liderbs.modelo.UsersId;
 
 import org.hibernate.Query;
 import org.hibernate.SessionFactory;
@@ -43,7 +42,7 @@ import javax.annotation.Resource;
  */
 @Scope("singleton")
 @Repository("UsersDAO")
-public class UsersDAO extends HibernateDaoImpl<Users, Long>
+public class UsersDAO extends HibernateDaoImpl<Users, Integer>
     implements IUsersDAO {
     private static final Logger log = LoggerFactory.getLogger(UsersDAO.class);
     @Resource
